fix 5/19/2024

+++ AdminApp/Controllers/ThuonghieuController.cs
@@ -7,30 +7,67 @@ namespace AdminApp.Controllers
 {
     public class ThuonghieuController : Controller
     {
-      
-            public HttpClient httpClient { get; set; }
-            public ThuonghieuController(HttpClient httpClient)
-            {
-                this.httpClient = httpClient;
-            }
-            public async Task<IActionResult> GetAllTH()
+
+        public HttpClient httpClient { get; set; }
+        public ThuonghieuController(HttpClient httpClient)
+        {
+            this.httpClient = httpClient;
+        }
+        public async Task<IActionResult> GetAllTH()
+        {
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
                 var th = JsonConvert.DeserializeObject<List<ThuongHieu>>(await (await httpClient.GetAsync("https://localhost:7050/api/ThuongHieu")).Content.ReadAsStringAsync());
                 return View(th);
             }
-            public async Task<IActionResult> DetailTH(Guid id)
+            else
             {
 
+                return RedirectToAction("Login", "Home");
+            }
+        }
+        public async Task<IActionResult> DetailTH(Guid id)
+        {
+
+           
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
+            {
                 var th = (JsonConvert.DeserializeObject<List<ThuongHieu>>(await (await httpClient.GetAsync("https://localhost:7050/api/ThuongHieu")).Content.ReadAsStringAsync())).FirstOrDefault(x => x.Guid == id);
                 return View(th);
-
             }
-            public ActionResult CreateTH()
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+        }
+        public ActionResult CreateTH()
+        {
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
                 return View();
             }
-            [HttpPost]
-            public async Task<IActionResult> CreateTH(ThuongHieu Th)
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+            
+        }
+        [HttpPost]
+        public async Task<IActionResult> CreateTH(ThuongHieu Th)
+        {
+           
+
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
                 // Serialize the Loai object to JSON
                 var jsonContent = JsonConvert.SerializeObject(Th);
@@ -50,30 +87,52 @@ namespace AdminApp.Controllers
                     return BadRequest(response);
                 }
             }
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+
+        }
 
         public async Task<IActionResult> DeleteTH(Guid id)
         {
-            var response = await httpClient.DeleteAsync($"https://localhost:7050/api/ThuongHieu/{id}");
-            if (response.IsSuccessStatusCode)
+         
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
-                // Xóa thành công
-                return RedirectToAction("GetAllTH");
+                var response = await httpClient.DeleteAsync($"https://localhost:7050/api/ThuongHieu/{id}");
+                if (response.IsSuccessStatusCode)
+                {
+                    // Xóa thành công
+                    return RedirectToAction("GetAllTH");
+                }
+                else
+                {
+                    // Xóa không thành công, in ra kết quả để debug
+                    var responseContent = await response.Content.ReadAsStringAsync();
+                    // Log hoặc in ra lỗi để kiểm tra
+                    Console.WriteLine($"Error deleting Loai: {responseContent}");
+
+                    // Thêm thông báo lỗi vào ViewData hoặc TempData để hiển thị trong View
+                    TempData["ErrorMessage"] = "Không thể xóa loại. Vui lòng thử lại.";
+
+                    return RedirectToAction("GetAllTH");
+                }
             }
             else
             {
-                // Xóa không thành công, in ra kết quả để debug
-                var responseContent = await response.Content.ReadAsStringAsync();
-                // Log hoặc in ra lỗi để kiểm tra
-                Console.WriteLine($"Error deleting Loai: {responseContent}");
 
-                // Thêm thông báo lỗi vào ViewData hoặc TempData để hiển thị trong View
-                TempData["ErrorMessage"] = "Không thể xóa loại. Vui lòng thử lại.";
-
-                return RedirectToAction("GetAllTH");
+                return RedirectToAction("Login", "Home");
             }
         }
 
         public async Task<IActionResult> EditTH(Guid Id)
+        {
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
                 var response = await httpClient.GetAsync($"https://localhost:7050/api/ThuongHieu/{Id}");
 
@@ -96,10 +155,21 @@ namespace AdminApp.Controllers
                     return NotFound();
                 }
             }
+            else
+            {
 
-            [HttpPost]
-            [ValidateAntiForgeryToken]
-            public async Task<IActionResult> EditTH(ThuongHieu th)
+                return RedirectToAction("Login", "Home");
+            }
+
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> EditTH(ThuongHieu th)
+        {          
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
                 var apiUrl = $"https://localhost:7050/api/ThuongHieu/{th.Guid}";
 
@@ -117,8 +187,15 @@ namespace AdminApp.Controllers
                     return BadRequest();
                 }
             }
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
 
         }
+
     }
+}
